feat: export the svelte/server types that the docs already reference
