+++ packages/svelte/src/internal/server/types.d.ts
@@ -15,8 +15,6 @@ export interface SSRContext {
 	element?: Element;
 }
 
-export type Csp = { nonce?: string; hash?: boolean };
-
 export interface HydratableLookupEntry {
 	value: unknown;
 	serialized: string;
@@ -34,19 +32,3 @@ export interface HydratableContext {
 export interface RenderContext {
 	hydratable: HydratableContext;
 }
-
-export type Sha256Source = `sha256-${string}`;
-
-export interface SyncRenderOutput {
-	/** HTML that goes into the `<head>` */
-	head: string;
-	/** @deprecated use `body` instead */
-	html: string;
-	/** HTML that goes somewhere into the `<body>` */
-	body: string;
-	hashes: {
-		script: Sha256Source[];
-	};
-}
-
-export type RenderOutput = SyncRenderOutput & PromiseLike<SyncRenderOutput>;
