+++ packages/svelte/src/server/index.d.ts
@@ -1,5 +1,7 @@
-import type { Csp, RenderOutput } from '#server';
+import type { Csp, RenderOutput } from './public.js';
 import type { ComponentProps, Component, SvelteComponent, ComponentType } from 'svelte';
+
+export type { Csp, RenderOutput, SyncRenderOutput, Sha256Source } from './public.js';
 
 /**
  * Only available on the server and when compiling with the `server` option.
